+++ tooling/api/src/webview.ts
@@ -417,15 +417,15 @@ class Webview {
       )
     }
 
+    const value = {} as Record<string, unknown>
+    value[`${size.type}`] = {
+      width: size.width,
+      height: size.height
+    }
+
     return invoke('plugin:webview|set_webview_size', {
       label: this.label,
-      value: {
-        type: size.type,
-        data: {
-          width: size.width,
-          height: size.height
-        }
-      }
+      value
     })
   }
 
@@ -452,15 +452,15 @@ class Webview {
       )
     }
 
+    const value = {} as Record<string, unknown>
+    value[`${position.type}`] = {
+      x: position.x,
+      y: position.y
+    }
+
     return invoke('plugin:webview|set_webview_position', {
       label: this.label,
-      value: {
-        type: position.type,
-        data: {
-          x: position.x,
-          y: position.y
-        }
-      }
+      value
     })
   }
 
@@ -477,6 +477,23 @@ class Webview {
   async setFocus(): Promise<void> {
     return invoke('plugin:webview|set_webview_focus', {
       label: this.label
+    })
+  }
+
+  /**
+   * Set webview zoom level.
+   * @example
+   * ```typescript
+   * import { getCurrent } from '@tauri-apps/api/webview';
+   * await getCurrent().setZoom(1.5);
+   * ```
+   *
+   * @returns A promise indicating the success or failure of the operation.
+   */
+  async setZoom(scaleFactor: number): Promise<void> {
+    return invoke('plugin:webview|set_webview_zoom', {
+      label: this.label,
+      value: scaleFactor
     })
   }
 
@@ -649,7 +666,15 @@ interface WebviewOptions {
    * */
   proxyUrl?: string
   /**
-   * Whether page zooming by hotkeys or gestures is enabled **Windows Only**
+   * Whether page zooming by hotkeys is enabled
+   *
+   * ## Platform-specific:
+   *
+   * - **Windows**: Controls WebView2's [`IsZoomControlEnabled`](https://learn.microsoft.com/en-us/microsoft-edge/webview2/reference/winrt/microsoft_web_webview2_core/corewebview2settings?view=webview2-winrt-1.0.2420.47#iszoomcontrolenabled) setting.
+   * - **MacOS / Linux**: Injects a polyfill that zooms in and out with `ctrl/command` + `-/=`,
+   * 20% in each step, ranging from 20% to 1000%. Requires `webview:allow-set-webview-zoom` permission
+   *
+   * - **Android / iOS**: Unsupported.
    */
   zoomHotkeysEnabled?: boolean
 }