+++ core/tauri-plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## \[2.0.0-beta.13]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.0.0-beta.13`
+
+## \[2.0.0-beta.12]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.0.0-beta.12`
+
 ## \[2.0.0-beta.11]
 
 ### Dependencies